Install Nginx, configure webservices (untested*)
Installed Nginx (as a reverse proxy), PHP-FPM (for Laravel app and
beanstalk-console) and Supervisor (to manage redis-commander).

Need to move paths and things to vars in the playbook rather than
hard-coding them.

* My Virtualbox now decided it doesn't like shared folders after a vagrant
reload. Only known solution is to destroy the vm and re-up (which on my
machine takes about 30 minutes). VMWare is looking cheaper and cheaper...